I have a feeling this will help increase performance

+++ app/src/main/jni/desmume/src/android/main.cpp
@@ -591,7 +591,7 @@ void JNI(restoreState, int slot)
 
 void loadSettings(JNIEnv* env)
 {
-	CommonSettings.num_cores = sysconf( _SC_NPROCESSORS_ONLN );
+	CommonSettings.num_cores = static_cast<int>(sysconf(_SC_NPROCESSORS_ONLN ));
 	LOGI("%i cores detected", CommonSettings.num_cores); 
 	CommonSettings.cheatsDisable = GetPrivateProfileBool(env,"General", "cheatsDisable", false, IniName);
 	CommonSettings.autodetectBackupMethod = GetPrivateProfileInt(env,"General", "autoDetectMethod", 0, IniName);
